backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
Pillow
python-dotenv
python-multipart
